requirements.txt
numpy<2.0.0
pyradiomics
platipy
PyYAML
pika
psycopg2-binary
